requirements.txt
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain_openai
openai
Ollama
langchain-groq
groq
faiss-cpu
Ollama
langchain_google_genai
chromadb
langchain-huggingface
langchain_chroma
pypdf
langchain-text-splitters